pyproject.toml
[project]
name = "docagent"
version = "0.1.0"
description = "Agente RAG local com Qwen2.5 + LangChain"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.11.4",
    "fastapi>=0.135.1",
    "sse-starlette>=3.3.3",
    "streamlit>=1.55.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "mcp>=1.26.0",
    "langchain-mcp-adapters>=0.2.2",
    "asyncpg>=0.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docagent"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
